Backport 1.1.0 features into 1.0 branch
This commit changes the state of the code in the 1.0 branch to include all of the changes that are currently in MASTER (which is at 1.2.0-SNAPSHOT), but excludes features that are not compatible with Openfire 4.2.3 (which seems to be limited to the backup functionality).

The intent of this commit is to release a 1.0.x version that's compatible with Openfire 4.2.3.